Better structure
- Add cert in config/application.rb
  Must match the cert of the registry
