+++ engines/registry/app/models/catalog_client.rb
