+++ engines/registry/app/models/concerns/registry_client.rb
@@ -1,0 +1,6 @@
+module RegistryClient
+  extend ActiveSupport::Concern
+  attr_reader :account
+  attr_reader :systems
+
+end
