+++ engines/registry/config/routes.rb
@@ -1,2 +1,3 @@
 Registry::Engine.routes.draw do
+  get 'authorize', to: 'registry#authorize'
 end
